requirements.txt
Flask
Flask-RESTful
Flask-PluginKit>=2.1.0
SpliceURL>=1.2
requests
setproctitle
gevent
gunicorn
enum34
pycrypto
psutil
docker>=2.0.0
supervisor
redis